# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"SentinelX","version":"1.0.0"},"paths":{"/api/auth/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/users/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_auth_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/upload-document":{"post":{"summary":"Upload Document","operationId":"upload_document_api_upload_document_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_upload_document_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyze-document":{"post":{"summary":"Analyze Document Api","operationId":"analyze_document_api_api_analyze_document_post","parameters":[{"name":"doc_id","in":"query","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dashboard":{"get":{"summary":"Dashboard Api","operationId":"dashboard_api_api_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}}}},"/api/tasks":{"get":{"summary":"Tasks Api","operationId":"tasks_api_api_tasks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListResponse"}}}}}}},"/api/alerts":{"get":{"summary":"Alerts Api","operationId":"alerts_api_api_alerts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertListResponse"}}}}}}},"/api/agent-logs":{"get":{"summary":"Agent Logs Api","operationId":"agent_logs_api_api_agent_logs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentLogListResponse"}}}}}}},"/api/audit-trail":{"get":{"summary":"Audit Trail Api","operationId":"audit_trail_api_api_audit_trail_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditTrailResponse"}}}}}}},"/api/chat":{"post":{"summary":"Chat Api","operationId":"chat_api_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/knowledge-graph/{document_id}":{"get":{"summary":"Knowledge Graph Api","operationId":"knowledge_graph_api_api_knowledge_graph__document_id__get","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowledgeGraphResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/notify":{"post":{"summary":"Notify","operationId":"notify_api_notify_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentLog":{"properties":{"id":{"type":"string","title":"Id"},"agent":{"type":"string","title":"Agent"},"action":{"type":"string","title":"Action"},"timestamp":{"type":"string","title":"Timestamp"}},"type":"object","required":["id","agent","action","timestamp"],"title":"AgentLog"},"AgentLogListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/AgentLog"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"AgentLogListResponse"},"AlertItem":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"severity":{"type":"string","title":"Severity"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","title","severity","created_at"],"title":"AlertItem"},"AlertListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/AlertItem"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"AlertListResponse"},"AnalysisResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"summary":{"type":"string","title":"Summary"},"compliance_score":{"type":"integer","title":"Compliance Score"},"risk_score":{"type":"integer","title":"Risk Score"},"maps":{"items":{"$ref":"#/components/schemas/MapItem"},"type":"array","title":"Maps"},"risks":{"items":{"$ref":"#/components/schemas/RiskItem"},"type":"array","title":"Risks"},"departments":{"items":{"type":"string"},"type":"array","title":"Departments"},"executive_insights":{"type":"string","title":"Executive Insights"},"agent_reasoning":{"items":{"type":"string"},"type":"array","title":"Agent Reasoning"},"sources":{"items":{"$ref":"#/components/schemas/SourceCitation"},"type":"array","title":"Sources","default":[]}},"type":"object","required":["document_id","summary","compliance_score","risk_score","maps","risks","departments","executive_insights","agent_reasoning"],"title":"AnalysisResponse"},"AuditEvent":{"properties":{"id":{"type":"string","title":"Id"},"event":{"type":"string","title":"Event"},"timestamp":{"type":"string","title":"Timestamp"}},"type":"object","required":["id","event","timestamp"],"title":"AuditEvent"},"AuditTrailResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/AuditEvent"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"AuditTrailResponse"},"Body_login_for_access_token_api_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_auth_token_post"},"Body_upload_document_api_upload_document_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_upload_document_post"},"ChatRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["document_id","message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"reply":{"type":"string","title":"Reply"},"sources":{"items":{"$ref":"#/components/schemas/SourceCitation"},"type":"array","title":"Sources","default":[]},"debug":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Debug"}},"type":"object","required":["reply"],"title":"ChatResponse"},"DashboardResponse":{"properties":{"compliance_score":{"type":"integer","title":"Compliance Score"},"total_documents":{"type":"integer","title":"Total Documents"},"pending_actions":{"type":"integer","title":"Pending Actions"},"risk_alerts":{"type":"integer","title":"Risk Alerts"},"recent_uploads":{"items":{"type":"string"},"type":"array","title":"Recent Uploads"},"trend":{"items":{"type":"integer"},"type":"array","title":"Trend"}},"type":"object","required":["compliance_score","total_documents","pending_actions","risk_alerts","recent_uploads","trend"],"title":"DashboardResponse"},"GraphEdge":{"properties":{"source":{"type":"string","title":"Source"},"target":{"type":"string","title":"Target"},"label":{"type":"string","title":"Label"}},"type":"object","required":["source","target","label"],"title":"GraphEdge"},"GraphNode":{"properties":{"id":{"type":"string","title":"Id"},"label":{"type":"string","title":"Label"},"type":{"type":"string","title":"Type"}},"type":"object","required":["id","label","type"],"title":"GraphNode"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"KnowledgeGraphResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"nodes":{"items":{"$ref":"#/components/schemas/GraphNode"},"type":"array","title":"Nodes"},"edges":{"items":{"$ref":"#/components/schemas/GraphEdge"},"type":"array","title":"Edges"}},"type":"object","required":["document_id","nodes","edges"],"title":"KnowledgeGraphResponse"},"MapItem":{"properties":{"title":{"type":"string","title":"Title"},"department":{"type":"string","title":"Department"},"deadline":{"type":"string","title":"Deadline"},"severity":{"type":"string","title":"Severity"},"source_section":{"type":"string","title":"Source Section"}},"type":"object","required":["title","department","deadline","severity","source_section"],"title":"MapItem"},"RiskItem":{"properties":{"risk":{"type":"string","title":"Risk"},"severity":{"type":"string","title":"Severity"},"reason":{"type":"string","title":"Reason"},"source_section":{"type":"string","title":"Source Section"}},"type":"object","required":["risk","severity","reason","source_section"],"title":"RiskItem"},"SourceCitation":{"properties":{"section_title":{"type":"string","title":"Section Title"},"snippet":{"type":"string","title":"Snippet"},"score":{"type":"number","title":"Score"}},"type":"object","required":["section_title","snippet","score"],"title":"SourceCitation"},"TaskItem":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"department":{"type":"string","title":"Department"},"priority":{"type":"string","title":"Priority"},"status":{"type":"string","title":"Status"},"deadline":{"type":"string","title":"Deadline"}},"type":"object","required":["id","title","department","priority","status","deadline"],"title":"TaskItem"},"TaskListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/TaskItem"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"TaskListResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UploadResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"filename":{"type":"string","title":"Filename"},"pages":{"type":"integer","title":"Pages"},"message":{"type":"string","title":"Message"}},"type":"object","required":["document_id","filename","pages","message"],"title":"UploadResponse"},"User":{"properties":{"username":{"type":"string","title":"Username"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"role":{"type":"string","title":"Role"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","required":["username","role"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/token"}}}}}}